+++ src/client/index.ts
@@ -88,7 +88,7 @@ import { BaseRequest, BaseResponse } from '../models/methods/baseMethod'
 import autofill from '../sugar/autofill'
 import getBalances from '../sugar/balances'
 import getFee from '../sugar/fee'
-import getLedgerIndex from '../sugar/ledgerVersion'
+import getLedgerIndex from '../sugar/ledgerIndex'
 import getOrderbook from '../sugar/orderbook'
 import { submitTransaction, submitSignedTransaction } from '../sugar/submit'
 import { ensureClassicAddress } from '../sugar/utils'
